requirements.txt
# https://www.django-rest-framework.org/
djangorestframework

# https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html
djangorestframework-simplejwt

# https://pypi.org/project/psycopg2/
psycopg2

# https://django-environ.readthedocs.io/en/latest/
django-environ